My first commit on C preprocessing

+++ 0x00-hello_world/6-size.c
@@ -12,6 +12,6 @@ printf("Size of a char: %lu bytes(s)", (unsigned long)sizeof(char));
 printf("Size of an int: %lu bytes(s)", (unsigned long)sizeof(int));
 printf("Size of a long int: %lu bytes(s)", (unsigned long)sizeof(long int));
 printf("Size of a long long int: %lu bytes(s)", (unsigned long)sizeof(long long int));
-printf("Size of a float: %lu bytes(s)", sizeof(float));
+printf("Size of a float: %lu bytes(s)", (unsigned long)sizeof(float));
 return (0);
 }
